refactor(eng-analytics): co-locate scopes and shared fixtures
