+++ products/engineering_analytics/backend/presentation/views/suite_health.py
@@ -28,7 +28,10 @@ from products.engineering_analytics.backend.presentation.views._base import (
 )
 
 
-class TestHealthActionsMixin(EngineeringAnalyticsViewSetBase):
+class SuiteHealthActionsMixin(EngineeringAnalyticsViewSetBase):
+    READ_ACTIONS = ["flaky_tests", "broken_tests", "quarantine"]
+    WRITE_ACTIONS = ["quarantine_request"]
+
     @extend_schema(
         operation_id="engineering_analytics_flaky_tests",
         parameters=[
